Move Polars constraints tests to new file
